net, setup: Move networkCacheStoreFactory & podInterfaceCache
The two cache related objects are moved from the VirtualMachineController
to the network setup controller.

As a temporary step, preserve network setup & update tests as is, with
only adjustment for them to keep working.

Following changes will move most of the tests to the network package,
keeping in the VirtualMachineController tests only the check that the
setup and other network entry points are called from the main flow.
